refactored multiplication
moved multiplication in own source files and link them accordingly
moved randomInRange to AES_common.c
