+++ AES_common.h
@@ -9,6 +9,7 @@ extern const word8 inv_s[];
 word8 multiplicationX(word8 byte);
 word8 multiplicationXN(word8 byte, int n);
 
+int randomInRange(int min, int max);
 word8 inverseByteTransformation(word8 byte);
 
 void encryption(const word8 initialMessage[][4], const word8 initialKey[][4], word8 ciphertext[4][4]);
